Refactor component naming and macro dispatch logic
- Rename UI components to avoid conflicts (e.g., Button → button_)
- Update module exports to use new component names
- Modify gen_dispatch macro to require explicit object parameter
- Adjust dynamic.rs to pass component object to macro
- Remove unused static ID counters in dynamic.rs

+++ ui/src/libs/components/float.rs
@@ -8,7 +8,7 @@ use std::ops::Deref;
 use std::rc::Rc;
 
 #[component]
-pub fn Float(layout: Layout, children: Element) -> Element {
+pub fn float_(layout: Layout, children: Element) -> Element {
     let mut css = vec!["float", "f"];
     use_common_css(&mut css, &layout);
     let style = if let Some(attr) = &layout.attrs
